+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project will be documented in this file[^1].
 ### Changed
 - favourites navbar button to always show
 
+### Fixed
+- Vue initialization in admin
+
 ## [2.32.1] - 2021-10-14
 ### Fixed
 - public views for shared user collections