+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+
+## [2.33.0] - 2021-10-29
 ### Added
 - default e-mail template, style and config
 - e-mail message for created shared user collections